Adding multi columns -- Robin

+++ scripts/main.js
@@ -2,6 +2,8 @@
 
 var scheduler = {
     
+    columns: 3,
+    
     hourStart: 6,
     hourEnd: 22,
     
@@ -10,18 +12,15 @@ var scheduler = {
     slotMargin: 1,
     
     slots: {},
-    types: {
-        1: {ord: 1, name: 'Apple', color: '#FF8C00'},
-        2: {ord: 2, name: 'Orange', color: '#9932CC'},
-        3: {ord: 3, name: 'Pear', color: '#8B0000'},
-    },
+    types: {},
     
     init: function() {
         var that = this;
         
         this.setHoursHTML();
+        this.setColumnsHTML();
         
-        $('#slots').sortable({
+        $('#slots ul').sortable({
             placeholder: 'ui-state-highlight',
             cancel: '.empty',
             start: function(event, ui) {
@@ -30,12 +29,11 @@ var scheduler = {
         });
         
         if (!Object.keys(this.slots).length) {
-            
             this.createEmptySlots((this.hourEnd - this.hourStart + 1) * 3);
         }
         
-        this.setTypesHTML();
         this.setSlotsHTML();
+        this.setTypesHTML();
         
         //Add a new type
         $('#add_new_type').on('click', function() {
@@ -113,7 +111,11 @@ var scheduler = {
         $('#types .type').draggable({
             helper: function(event) {
                 var typeId = $(event.currentTarget).data('id');
-                return $('<div>', {text: that.types[typeId].name, class: 'test'});
+                return $('<div>', {
+                    text: that.types[typeId].name, 
+                    class: 'helper', 
+                    style: 'background-color:' + that.types[typeId].color
+                });
             },
             handle: '.drag',
             cursorAt: {top:10, left: 0}
@@ -121,10 +123,13 @@ var scheduler = {
         
         $('#types .type .close').on('click', function() {
             var typeId = $(this).parent().parent().data('id');
-            that.deleteType(typeId);
-            
-            that.setTypesHTML();
-            that.setSlotsHTML();
+            if (typeId && confirm('Are you sure you want to delete the type "' + that.types[typeId].name + '"?')) {
+                
+                that.deleteType(typeId);
+                
+                that.setTypesHTML();
+                that.setSlotsHTML();
+            }
         });
         
         $('#types .type input[type="color"]').on('change', function() {
@@ -137,7 +142,8 @@ var scheduler = {
     
     updateSlotType: function(slotId, typeId) {
         var slot = {
-            ord: this.slots[slotId].ord
+            ord: this.slots[slotId].ord,
+            columnId: this.slots[slotId].columnId
         };
         if (typeId) {
             slot.typeId = typeId;
@@ -175,11 +181,13 @@ var scheduler = {
         return slot;
     },
     
-    getSlots: function() {
+    getSlots: function(columnId) {
         var slots = [];
         for (var slotId in this.slots) {
-            var slot = this.getSlotMergeFields(slotId);
-            slots.push(slot);
+            if (this.slots[slotId].columnId == columnId) {
+                var slot = this.getSlotMergeFields(slotId);
+                slots.push(slot);
+            }
         }
         slots.sort(function(a, b) {
             return a.ord - b.ord;
@@ -187,7 +195,7 @@ var scheduler = {
         return slots;
     },
     
-    getSlotsHTML: function(slotId) {
+    getSlotsHTML: function(columnId, slotId) {
         var source = document.getElementById('slot-template').innerHTML;
         var template = Handlebars.compile(source);
         var html = '';
@@ -195,7 +203,7 @@ var scheduler = {
             var slot = this.getSlotMergeFields(slotId);
             html = template(slot);
         } else {
-            var slots = this.getSlots();
+            var slots = this.getSlots(columnId);
             for (var i = 0; i < slots.length; i++) {
                 html += template(slots[i]);
             }
@@ -207,9 +215,11 @@ var scheduler = {
         var that = this;
         
         //Save existing slot order
-        that.saveSlotOrder();
-        var html = this.getSlotsHTML();
-        $('#slots').html(html);
+        for (var i = 0; i < this.columns; i++) {
+            that.saveSlotOrder(i);
+            var html = this.getSlotsHTML(i);
+            $('#column_' + i).html(html);
+        }
         
         $('#slots .empty').droppable({
             accept: '.type',
@@ -235,9 +245,11 @@ var scheduler = {
             grid: [0, this.slotHeight + this.slotMargin + (this.slotBorder * 2)],
             start: function(event, ui) {
                 //Set max height
-                that.saveSlotOrder();
                 var slotId = $(this).data('id');
-                var slots = that.getSlots();
+                var slot = that.slots[slotId];
+                that.saveSlotOrder(slot.columnId);
+                
+                var slots = that.getSlots(slot.columnId);
                 var passed = false;
                 
                 emptySlotCount = that.slots[slotId].size - 1;
@@ -253,6 +265,7 @@ var scheduler = {
                         }
                     }
                 }
+                
                 var maxHeight = ((emptySlotCount) * (that.slotHeight + that.slotMargin + (that.slotBorder * 2))) + that.slotHeight;
                 $('#slots .item').resizable('option', 'maxHeight', maxHeight);
             },
@@ -275,21 +288,21 @@ var scheduler = {
                 if (newSize > oldSize) {
                     //When size is increased, remove the slot below
                     for (var j = oldSize; j < newSize; j++) {
-                        var slots = that.getSlots();
+                        var slots = that.getSlots(slot.columnId);
                         var passed = false;
-                        for (var i = 0; i < slots.length; i++) {
-                            if (slots[i].id == slotId) {
+                        for (var k = 0; k < slots.length; k++) {
+                            if (slots[k].id == slotId) {
                                 passed = true;
                             } else if (passed) {
-                                if (slots[i].empty) {
-                                    $('#slots li[data-id="' + slots[i].id + '"]').remove();
-                                    delete that.slots[slots[i].id];
+                                if (slots[k].empty) {
+                                    $('#slots li[data-id="' + slots[k].id + '"]').remove();
+                                    delete that.slots[slots[k].id];
                                     break;
                                 }
                             }
                         }
                     }
-                } else {
+                } else if (newSize < oldSize) {
                     //When size is decreased, add an empty slot below
                     var count = oldSize - newSize;
                     that.createEmptySlots(count, slotId);
@@ -304,27 +317,45 @@ var scheduler = {
         });
     },
     
+    //Either fills all columns with {count} slots
+    //Or inserts {count} slots under slot {slotId}
     createEmptySlots: function(count, slotId) {
-        var $el = slotId ? $('#slots li[data-id="' + slotId + '"]') : $('#slots');
-        for (var i = 0; i < count; i++) {
-            var ids = Object.keys(this.slots);
-            var id = ids.length ? Math.max(...ids.map(Number)) + 1 : 1;
-            this.slots[id] = {empty: true, ord: 0};
-            
-            var html = this.getSlotsHTML(id);
+        for (var i = 0; i < this.columns; i++) {
             if (slotId) {
-                $el.after(html);
+                var $el = $('#slots li[data-id="' + slotId + '"]');
+                var columnId = this.slots[slotId].columnId;
             } else {
-                $el.append(html);
+                var $el = $('#column_' + i);
+                var columnId = i;
             }
-            
-            this.saveSlotOrder();
+            for (var j = 0; j < count; j++) {
+                //Get next slot Id
+                var ids = Object.keys(this.slots);
+                var id = ids.length ? Math.max(...ids.map(Number)) + 1 : 1;
+                this.slots[id] = {empty: true, ord: 0, columnId: columnId};
+                
+                //Get HTML for empty slot
+                var html = this.getSlotsHTML(false, id);
+                
+                //Set HTML
+                if (slotId) {
+                    $el.after(html);
+                } else {
+                    $el.append(html);
+                }
+                
+                //Save the order of the slots (so I don't have to manually calculate ordinals)
+                this.saveSlotOrder(columnId);
+            }
+            if (slotId) {
+                break;
+            }
         }
     },
     
-    saveSlotOrder: function() {
+    saveSlotOrder: function(columnId) {
         var ord = 1;
-        var slotIds = $('#slots').sortable('toArray', {attribute: 'data-id'});        
+        var slotIds = $('#column_' + columnId).sortable('toArray', {attribute: 'data-id'});        
         for (var i = 0; i < slotIds.length; i++) {
             if (this.slots[slotIds[i]]) {
                 this.slots[slotIds[i]].ord = ord++;
@@ -338,6 +369,14 @@ var scheduler = {
             html += '<div class="hour">' + i + ':00</div>';
         }
         $('#hours').html(html);
+    },
+    
+    setColumnsHTML: function() {
+        var html = '';
+        for (var i = 0; i < this.columns; i++) {
+            html += '<ul id="column_' + i + '"></ul>';
+        }
+        $('#slots').html(html);
     }
     
 };
